Module 3 started: - a lot of changes related to hiberanate - move some entities to DB - implement Account and user paralel creation - implement refill method for account - refactor all id in models to be Long (Integer was initial ) - refill method on UI

+++ src/main/java/epamUniversity/model/DomainObject.java
@@ -1,17 +1,21 @@
 package epamUniversity.model;
 
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
 /**
  * Created by Andriy_Yarish on 1/3/2017.
  */
 public class DomainObject {
 
-    private Integer id;
+    private Long id;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
